fix(cli): sanitize image metadata before upload so agents can publish charts
The relay rejects images carrying metadata — `iCCP`, `eXIf`, `iTXt`, `zTXt`,
`pHYs` and any other non-allowlisted ancillary chunk. Desktop strips metadata
before hashing; `buzz-cli` never did, so `buzz upload file` returned

    422 media contains metadata or a non-canonical metadata channel

for anything a real producer emits. That is every macOS screenshot
(`iCCP`+`eXIf`+`iTXt`) and every matplotlib figure (`tEXt`+`pHYs`). Agents
could read images but not publish them: no charts, no plots, no diagrams.

The relay cannot fix this server-side. Blossom is content-addressed, so the
kind-24242 upload auth event commits to the SHA-256 of the bytes actually
sent — stripping has to happen client-side, before the hash.

Extract the sanitizer and the metadata validator into a new leaf crate,
`buzz-image`, and call it from the CLI before hashing. The two are halves of
one contract — every client must produce what the relay accepts — and they
drifted precisely because they lived in different crates with no shared test.
A round-trip test now pins them together over fixtures taken from real
producers, including the `tEXt`+`pHYs` pair matplotlib writes by default.

`buzz-image` is a leaf on purpose: `image`, `infer`, `thiserror` and nothing
else. `buzz-media` pulls axum and an S3 client, so hoisting the sanitizer
there would have linked a web framework into the CLI binary. Feature-gating
would not have helped — under `resolver = "2"` a shared normal dependency
unifies across workspace members, so `cargo build --workspace` would switch
the feature back on for the CLI.

Also collapses the snapshot-keyword allowlist, which existed as three
hardcoded copies across two crates. Adding a third snapshot kind would have
produced a file the producer wrote, the sanitizer preserved, and the relay
silently rejected.

Behaviour notes:

- Validate first, sanitize only on failure. Files the relay already accepts
  are uploaded untouched and keep their on-disk SHA-256; only files that
  would have 422ed are rewritten.
- Sanitizing is best-effort. A payload that sniffs as an image but cannot be
  decoded is uploaded as-is rather than rejected locally — the relay stays
  the authority, so this can never refuse an upload that previously worked.
  The two deliberate refusals (animated PNG/WebP with EXIF orientation or an
  ICC profile, which cannot be stripped without visibly changing the image)
  still propagate with their specific message.
- The still-image path decodes and re-encodes, which is lossy for JPEG and
  inflates PNGs by +11% to +65% (measured across 7 real macOS screenshots;
  the largest file inflates worst). JPEG re-encodes at the `image` crate
  default rather than the quality 95 used elsewhere in `media.rs`. Both are
  pre-existing Desktop behaviour that the CLI now inherits; validate-first
  limits them to files that would otherwise be rejected outright. Structural
  chunk stripping would make both lossless, as the GIF path already is —
  follow-up, not this change.
- Size caps and the 25-megapixel ceiling stay in `buzz-media`: they are
  server policy and map to 413, not 422. Status codes are pinned by test
  across the crate boundary.
