+++ crates/buzz-cli/Cargo.toml
@@ -45,6 +45,10 @@ chrono = { workspace = true }
 buzz-sdk = { workspace = true }
 buzz-core = { workspace = true }
 
+# Shared image sanitizer + metadata validator — the relay rejects images
+# carrying metadata, and content addressing means we must strip before hashing.
+buzz-image = { workspace = true }
+
 # Base64 encoding — NIP-98 event serialization for Authorization header
 base64 = "0.22"
