+++ crates/buzz-image/src/snapshot_png.rs
@@ -1,17 +1,11 @@
 //! Preservation of Buzz snapshot tEXt chunks through the upload sanitizer.
 //!
-//! Split out of `media.rs` to keep that file under the desktop line-size
-//! limit. Agent/team sharing embeds a manifest in a PNG `tEXt` chunk
+//! Agent/team sharing embeds a manifest in a PNG `tEXt` chunk
 //! (`buzz_agent_snapshot` / `buzz_team_snapshot`); the sanitizer's re-encode
 //! would destroy it and the relay would previously reject it. These helpers
 //! extract the chunk before the re-encode and re-inject it afterwards. The
-//! relay allowlists exactly these keywords in `buzz-media::validation` — the
-//! two lists must stay in sync.
-
-/// tEXt keywords that carry Buzz snapshot manifests (`.agent.png` /
-/// `.team.png`). These chunks are the product payload of agent/team sharing —
-/// they must survive the metadata strip.
-const PNG_SNAPSHOT_KEYWORDS: [&[u8]; 2] = [b"buzz_agent_snapshot", b"buzz_team_snapshot"];
+//! allowlist itself lives in `crate::validate` and is shared with the relay's
+//! validator, so there is exactly one list.
 
 /// Extract the raw bytes of the first Buzz snapshot tEXt chunk (length + type
 /// + payload + CRC) from a PNG, or `None` when absent/not a PNG.
@@ -19,7 +13,7 @@ const PNG_SNAPSHOT_KEYWORDS: [&[u8]; 2] = [b"buzz_agent_snapshot", b"buzz_team_s
 /// Walks the chunk structure directly instead of decoding the image so a
 /// malformed file simply yields `None` and falls through to the normal
 /// sanitize path.
-pub(crate) fn extract_snapshot_text_chunk(bytes: &[u8]) -> Option<Vec<u8>> {
+pub fn extract_snapshot_text_chunk(bytes: &[u8]) -> Option<Vec<u8>> {
     const SIG: &[u8] = b"\x89PNG\r\n\x1a\n";
     if !bytes.starts_with(SIG) {
         return None;
@@ -34,11 +28,13 @@ pub(crate) fn extract_snapshot_text_chunk(bytes: &[u8]) -> Option<Vec<u8>> {
         let kind = &bytes[i + 4..i + 8];
         if kind == b"tEXt" {
             let payload = &bytes[i + 8..i + 8 + len];
-            let is_snapshot = PNG_SNAPSHOT_KEYWORDS.iter().any(|keyword| {
-                payload.len() > keyword.len()
-                    && &payload[..keyword.len()] == *keyword
-                    && payload[keyword.len()] == 0
-            });
+            let is_snapshot = crate::validate::PNG_SNAPSHOT_KEYWORDS
+                .iter()
+                .any(|keyword| {
+                    payload.len() > keyword.len()
+                        && &payload[..keyword.len()] == *keyword
+                        && payload[keyword.len()] == 0
+                });
             if is_snapshot {
                 return Some(bytes[i..end].to_vec());
             }
@@ -56,7 +52,7 @@ pub(crate) fn extract_snapshot_text_chunk(bytes: &[u8]) -> Option<Vec<u8>> {
 /// the import path only exposes text chunks encountered before IDAT via
 /// `read_info()`. The chunk bytes carry their own CRC, which remains valid
 /// because the chunk content is unchanged.
-pub(crate) fn inject_snapshot_text_chunk(png: Vec<u8>, chunk: &[u8]) -> Result<Vec<u8>, String> {
+pub fn inject_snapshot_text_chunk(png: Vec<u8>, chunk: &[u8]) -> Result<Vec<u8>, String> {
     const SIG_LEN: usize = 8;
     // IHDR is always the first chunk of a well-formed PNG.
     if png.len() < SIG_LEN + 12 || &png[SIG_LEN + 4..SIG_LEN + 8] != b"IHDR" {
@@ -81,7 +77,7 @@ pub(crate) fn inject_snapshot_text_chunk(png: Vec<u8>, chunk: &[u8]) -> Result<V
 
 #[cfg(test)]
 mod tests {
-    use super::super::media::sanitize_image_for_upload;
+    use crate::sanitize_image_for_upload;
 
     #[test]
     fn test_sanitizer_preserves_agent_snapshot_text_chunk() {
@@ -140,89 +136,5 @@ mod tests {
         let decoder = png::Decoder::new(std::io::Cursor::new(&sanitized));
         let reader = decoder.read_info().unwrap();
         assert!(reader.info().uncompressed_latin1_text.is_empty());
-    }
-
-    #[test]
-    fn test_agent_snapshot_survives_full_share_pipeline() {
-        // Cross-contract regression for the production failure: a real
-        // encoded .agent.png must survive export → client sanitize →
-        // relay validation → import decode. Each seam is also covered by
-        // unit tests, but this proves the exact pipeline that broke.
-        use crate::managed_agents::agent_snapshot::{
-            decode_snapshot_png, encode_snapshot_png, AgentSnapshot, AgentSnapshotDefinition,
-            AgentSnapshotMemory, AgentSnapshotProfile, MemoryLevel,
-        };
-
-        let snapshot = AgentSnapshot {
-            format: "buzz-agent-snapshot".to_string(),
-            version: 1,
-            definition: AgentSnapshotDefinition {
-                name: "Tree Trunks".to_string(),
-                source_is_builtin: false,
-                system_prompt: Some("You are a helpful agent.".to_string()),
-                runtime: Some("goose".to_string()),
-                model: None,
-                provider: None,
-                parallelism: Some(1),
-                respond_to: None,
-                respond_to_allowlist: vec![],
-                idle_timeout_seconds: None,
-                max_turn_duration_seconds: None,
-                name_pool: vec![],
-            },
-            profile: AgentSnapshotProfile {
-                display_name: "Tree Trunks".to_string(),
-                about: Some("Shared agent".to_string()),
-                avatar_data_url: None,
-                avatar_url: None,
-            },
-            memory: AgentSnapshotMemory {
-                level: MemoryLevel::None,
-                entries: vec![],
-            },
-        };
-
-        // Export with a real avatar body (what the share flow uploads).
-        let avatar = image::DynamicImage::ImageRgba8(image::RgbaImage::from_pixel(
-            4,
-            4,
-            image::Rgba([200, 120, 40, 255]),
-        ));
-        let mut avatar_png = std::io::Cursor::new(Vec::new());
-        avatar
-            .write_to(&mut avatar_png, image::ImageFormat::Png)
-            .unwrap();
-        let exported = encode_snapshot_png(&snapshot, Some(avatar_png.get_ref())).unwrap();
-
-        // Client upload path.
-        let sanitized = sanitize_image_for_upload(exported, "image/png").unwrap();
-
-        // Relay ingest path.
-        let relay_config = buzz_media_pkg::MediaConfig {
-            s3_endpoint: String::new(),
-            s3_access_key: String::new(),
-            s3_secret_key: String::new(),
-            s3_bucket: String::new(),
-            s3_region: "us-east-1".to_string(),
-            s3_addressing_style: buzz_media_pkg::S3AddressingStyle::Path,
-            max_image_bytes: 50 * 1024 * 1024,
-            max_gif_bytes: 10 * 1024 * 1024,
-            max_video_bytes: 524_288_000,
-            max_file_bytes: 104_857_600,
-            public_base_url: String::new(),
-            upload_records_enabled: false,
-            upload_ip_header: None,
-            upload_port_header: None,
-        };
-        assert_eq!(
-            buzz_media_pkg::validation::validate_content(&sanitized, &relay_config)
-                .expect("relay rejected a sanitized agent snapshot PNG"),
-            "image/png"
-        );
-
-        // Recipient import path.
-        let imported = decode_snapshot_png(&sanitized)
-            .expect("import failed on a PNG that passed sanitize + relay validation");
-        assert_eq!(imported, snapshot);
     }
 }
