+++ desktop/src-tauri/Cargo.toml
@@ -68,6 +68,8 @@ keyring = { version = "3.6.3", default-features = false, features = ["windows-na
 user-idle = { version = "0.6", default-features = false }
 
 [dependencies]
+# Shared image sanitizer + metadata validator (also used by the relay and CLI).
+buzz-image = { path = "../../crates/buzz-image" }
 atomic-write-file = "0.3"
 anyhow = "1"
 dirs = "6"
