+++ desktop/src-tauri/src/commands/media.rs
@@ -153,151 +153,9 @@ pub(crate) fn sanitize_filename(name: &str) -> String {
     }
 }
 
-/// Return true when a PNG/WebP payload declares animation.
-///
-/// Animated payloads use structural sanitizers so frame timing, looping, and
-/// disposal semantics are preserved without flattening the image. The relay's
-/// validator remains the final authority for the sanitized container.
-fn is_animated_image(body: &[u8], mime: &str) -> bool {
-    match mime {
-        "image/png" if body.starts_with(b"\x89PNG\r\n\x1a\n") => {
-            let mut offset = 8usize;
-            while offset.checked_add(12).is_some_and(|end| end <= body.len()) {
-                let length = u32::from_be_bytes([
-                    body[offset],
-                    body[offset + 1],
-                    body[offset + 2],
-                    body[offset + 3],
-                ]) as usize;
-                let Some(end) = offset.checked_add(12).and_then(|v| v.checked_add(length)) else {
-                    return false;
-                };
-                if end > body.len() {
-                    return false;
-                }
-                if &body[offset + 4..offset + 8] == b"acTL" {
-                    return true;
-                }
-                offset = end;
-            }
-            false
-        }
-        "image/webp"
-            if body.len() >= 12 && body.starts_with(b"RIFF") && &body[8..12] == b"WEBP" =>
-        {
-            let mut offset = 12usize;
-            while offset.checked_add(8).is_some_and(|end| end <= body.len()) {
-                let chunk = &body[offset..offset + 4];
-                if chunk == b"ANIM" || chunk == b"ANMF" {
-                    return true;
-                }
-                let length = u32::from_le_bytes([
-                    body[offset + 4],
-                    body[offset + 5],
-                    body[offset + 6],
-                    body[offset + 7],
-                ]) as usize;
-                let padded = length.checked_add(length & 1);
-                let Some(end) = padded.and_then(|v| offset.checked_add(8 + v)) else {
-                    return false;
-                };
-                if end > body.len() {
-                    return false;
-                }
-                offset = end;
-            }
-            false
-        }
-        _ => false,
-    }
-}
-
-pub(crate) fn sanitize_image_for_upload(body: Vec<u8>, mime: &str) -> Result<Vec<u8>, String> {
-    let format = match mime {
-        "image/jpeg" => image::ImageFormat::Jpeg,
-        "image/png" => image::ImageFormat::Png,
-        "image/webp" => image::ImageFormat::WebP,
-        // GIF is never re-encoded (that would destroy animation timing);
-        // metadata extensions are stripped structurally instead. Unparseable
-        // payloads pass through — the relay's validator is the authority.
-        "image/gif" => {
-            let stripped = super::media_gif::strip_gif_metadata(&body);
-            return Ok(stripped.unwrap_or(body));
-        }
-        _ => return Ok(body),
-    };
-
-    if is_animated_image(&body, mime) {
-        let oriented_format = match mime {
-            "image/png" if super::media_animated::animated_png_uses_exif_orientation(&body) => {
-                Some("PNG")
-            }
-            "image/webp" if super::media_animated::animated_webp_uses_exif_orientation(&body) => {
-                Some("WebP")
-            }
-            _ => None,
-        };
-        if let Some(format) = oriented_format {
-            return Err(format!(
-                "animated {format} with EXIF orientation cannot be uploaded without changing its appearance"
-            ));
-        }
-        let color_profile_format = match mime {
-            "image/png" if super::media_animated::animated_png_uses_icc_profile(&body) => {
-                Some("PNG")
-            }
-            "image/webp" if super::media_animated::animated_webp_uses_icc_profile(&body) => {
-                Some("WebP")
-            }
-            _ => None,
-        };
-        if let Some(format) = color_profile_format {
-            return Err(format!(
-                "animated {format} with an ICC profile cannot be uploaded without changing its colors"
-            ));
-        }
-        let stripped = match mime {
-            "image/png" => super::media_animated::strip_animated_png_metadata(&body),
-            "image/webp" => super::media_animated::strip_animated_webp_metadata(&body),
-            _ => None,
-        };
-        return Ok(stripped.unwrap_or(body));
-    }
-
-    // Agent/team snapshot PNGs carry their manifest in a tEXt chunk that the
-    // re-encode below would destroy. Pull it out first and re-inject it after
-    // sanitizing — all other metadata is still stripped, and the relay
-    // allowlists exactly this chunk.
-    let snapshot_chunk = if format == image::ImageFormat::Png {
-        super::media_snapshot_png::extract_snapshot_text_chunk(&body)
-    } else {
-        None
-    };
-
-    use image::ImageDecoder;
-    let reader = image::ImageReader::with_format(std::io::Cursor::new(&body), format);
-    let mut decoder = reader
-        .into_decoder()
-        .map_err(|_| "failed to decode image for metadata removal".to_string())?;
-    decoder
-        .set_limits(image::Limits::default())
-        .map_err(|_| "image exceeds safe decoding limits".to_string())?;
-    let orientation = decoder
-        .orientation()
-        .map_err(|_| "failed to read image orientation".to_string())?;
-    let mut image = image::DynamicImage::from_decoder(decoder)
-        .map_err(|_| "failed to decode image for metadata removal".to_string())?;
-    image.apply_orientation(orientation);
-    let mut output = std::io::Cursor::new(Vec::new());
-    image
-        .write_to(&mut output, format)
-        .map_err(|_| "failed to encode image without metadata".to_string())?;
-    let sanitized = output.into_inner();
-    match snapshot_chunk {
-        Some(chunk) => super::media_snapshot_png::inject_snapshot_text_chunk(sanitized, &chunk),
-        None => Ok(sanitized),
-    }
-}
+/// Metadata stripping lives in `buzz-image`, shared with the relay's validator
+/// and the CLI. Re-exported here so the existing call sites read unchanged.
+pub(crate) use buzz_image::sanitize_image_for_upload;
 
 pub(crate) fn detect_and_validate_mime(body: &[u8]) -> Result<String, String> {
     let mime = infer::get(body)
@@ -931,22 +789,6 @@ mod tests {
             image::load_from_memory_with_format(&sanitized, image::ImageFormat::Jpeg).unwrap();
         assert_eq!((decoded.width(), decoded.height()), (3, 2));
         assert!(!sanitized.windows(6).any(|bytes| bytes == b"Exif\0\0"));
-    }
-
-    #[test]
-    fn test_animated_png_and_webp_are_not_flattened() {
-        let mut apng = b"\x89PNG\r\n\x1a\n".to_vec();
-        apng.extend_from_slice(&8u32.to_be_bytes());
-        apng.extend_from_slice(b"acTL");
-        apng.extend_from_slice(&[0; 8]);
-        apng.extend_from_slice(&[0; 4]);
-        assert!(is_animated_image(&apng, "image/png"));
-        assert!(sanitize_image_for_upload(apng, "image/png").is_ok());
-
-        let mut webp = b"RIFF\x0c\0\0\0WEBPANIM".to_vec();
-        webp.extend_from_slice(&0u32.to_le_bytes());
-        assert!(is_animated_image(&webp, "image/webp"));
-        assert!(sanitize_image_for_upload(webp, "image/webp").is_ok());
     }
 
     #[test]
