+++ desktop/src-tauri/src/commands/mod.rs
@@ -26,11 +26,8 @@ mod join_policy;
 mod legacy_storage;
 mod link_preview;
 pub(crate) mod media;
-mod media_animated;
 mod media_download;
-mod media_gif;
 mod media_raw;
-mod media_snapshot_png;
 mod media_transcode;
 mod media_upload_progress;
 #[cfg(feature = "mesh-llm")]
